Rough prototype commit.
This installs the CMAKE script modules that define a CSC (C SHARP, C#)
compiler. The basic backbone here is the construction of a csc adapter
compiler to help mimic a "compile and link". This allows us to ride on top
of the existing Makefile structure for C/C++, with minimal changes to
the CMAKE source itself.

The adapter will have a 'compile' phase that records a '.backtrace'
file. In the .backtrace file there is a reference to the original CS file.
The .backtrace files are like object files in a C/C++ compile. The adapter
will be involved in the link phase, where it will get a list of .backtrace
object files, read the original source names and build the final assembly
based on all the original source (.cs) files.

Currently this patch is very window's centric, but I hope to start porting
the makefile concept to Mono/Linux shortly.

Currently does not support generation of csprojs.

Code is from an old defunk branch, so the port over to this
development stream may not work just quite yet.