Add shared_library support to csc_adapter.
Actually the LinkExecutable had the bulk of the support already, and was
generic enough to be manipulated via the CMAKE information. So this patch
really renames the LinkExecutable task as "Linker", and repurposes it for
both the EXE and DLL.

Next step is to compile and link a prexisting assembly.